+++ azure-linux/main.tf
@@ -2,11 +2,11 @@ terraform {
   required_providers {
     coder = {
       source  = "coder/coder"
-      version = "0.5.2"
+      version = "0.5.3"
     }
     azurerm = {
       source  = "hashicorp/azurerm"
-      version = "=3.0.0"
+      version = "=3.28.0"
     }
   }
 }
@@ -114,6 +114,14 @@ resource "coder_app" "code-server" {
   name     = "VS Code"
   url      = "http://localhost:13337/?folder=/home/${lower(substr(data.coder_workspace.me.owner, 0, 32))}"
   icon     = "/icon/code.svg"
+  subdomain = false
+  share     = "owner"
+
+  healthcheck {
+    url       = "http://localhost:13337/healthz"
+    interval  = 6
+    threshold = 20
+  } 
 }
 
 
@@ -246,19 +254,103 @@ resource "azurerm_virtual_machine_data_disk_attachment" "home" {
 resource "coder_metadata" "workspace_info" {
   count       = data.coder_workspace.me.start_count
   resource_id = azurerm_linux_virtual_machine.main[0].id
-
+  icon = "/icon/memory.svg"
   item {
-    key   = "type"
+    key   = "instance type"
     value = azurerm_linux_virtual_machine.main[0].size
   }
+  item {
+    key   = "location"
+    value = var.location
+  }  
+  item {
+    key   = "image"
+    value = azurerm_linux_virtual_machine.main[0].source_image_reference[0].offer
+  }  
+  item {
+    key   = "ip address"
+    value = azurerm_public_ip.main.ip_address
+  }    
 }
 
 resource "coder_metadata" "home_info" {
   resource_id = azurerm_managed_disk.home.id
-
+  icon = "/icon/database.svg"
   item {
     key   = "size"
     value = "${var.home_size} GiB"
   }
 
+}
+
+
+resource "coder_metadata" "hide_azurerm_resource_group" {
+  count = data.coder_workspace.me.start_count
+  resource_id = azurerm_resource_group.main.id
+  hide = true
+  item {
+    key = "name"
+    value = azurerm_resource_group.main.name
+  }  
+}
+
+
+resource "coder_metadata" "hide_azurerm_public_ip" {
+  count = data.coder_workspace.me.start_count
+  resource_id = azurerm_public_ip.main.id
+  hide = true
+  item {
+    key = "name"
+    value = azurerm_public_ip.main.name
+  }  
+}
+
+resource "coder_metadata" "hide_azurerm_virtual_network" {
+  count = data.coder_workspace.me.start_count
+  resource_id = azurerm_virtual_network.main.id
+  hide = true
+  item {
+    key = "name"
+    value = azurerm_virtual_network.main.name
+  }  
+}
+
+resource "coder_metadata" "hide_azurerm_subnet" {
+  count = data.coder_workspace.me.start_count
+  resource_id = azurerm_subnet.internal.id
+  hide = true
+  item {
+    key = "name"
+    value = azurerm_subnet.internal.name
+  }  
+}
+
+resource "coder_metadata" "hide_azurerm_network_interface" {
+  count = data.coder_workspace.me.start_count
+  resource_id = azurerm_network_interface.main.id
+  hide = true
+  item {
+    key = "name"
+    value = azurerm_network_interface.main.name
+  }  
+}
+
+resource "coder_metadata" "hide_tls_private_key" {
+  count = data.coder_workspace.me.start_count
+  resource_id = tls_private_key.dummy.id
+  hide = true
+  item {
+    key = "name"
+    value = tls_private_key.dummy.algorithm
+  }  
+}
+
+resource "coder_metadata" "hide_azurerm_virtual_machine_data_disk_attachment" {
+  count = data.coder_workspace.me.start_count
+  resource_id = azurerm_virtual_machine_data_disk_attachment.home[0].id
+  hide = true
+  item {
+    key = "name"
+    value = azurerm_virtual_machine_data_disk_attachment.home[0].managed_disk_id
+  }  
 }