+++ go/pod-goland-image/main.tf
@@ -2,7 +2,7 @@ terraform {
   required_providers {
     coder = {
       source  = "coder/coder"
-      version = "~> 0.4.15"
+      version = "~> 0.5.3"
     }
     kubernetes = {
       source  = "hashicorp/kubernetes"
@@ -135,6 +135,14 @@ resource "coder_app" "code-server" {
   name     = "VS Code"
   icon     = "/icon/code.svg"
   url      = "http://localhost:13337"
+  subdomain = false
+  share     = "owner"
+
+  healthcheck {
+    url       = "http://localhost:13337/healthz"
+    interval  = 3
+    threshold = 10
+  }   
 }
 
 resource "coder_app" "goland" {
@@ -142,6 +150,14 @@ resource "coder_app" "goland" {
   name     = "GoLand"
   icon     = "/icon/goland.svg"
   url      = "http://localhost:9001"
+  subdomain = false
+  share     = "owner"
+
+  healthcheck {
+    url       = "http://localhost:9001/healthz"
+    interval  = 3
+    threshold = 10
+  }    
 }
 
 resource "kubernetes_pod" "main" {